backend/requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
Flask-MySQLdb==2.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.8
pillow==12.1.1
python-dotenv==1.2.2
requests==2.32.5
urllib3==2.6.3
Werkzeug==3.1.6
